requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
langsmith==0.4.37
openai==2.2.0
orjson==3.11.3
packaging==25.0
pydantic==2.12.0
pydantic_core==2.41.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
zstandard==0.25.0
python-dotenv==1.0.1

